test(remote_exec): add failing test with no-remote-exec + disk_cache
This case covers a case where:
- target //a:first creates a file first.txt (locally, by
  tagging as "no-remote" with "--modify_execution_info").
- target //a:second copies first.txt to second.txt, remotely.

When --disk_cache is set, the first FindMissingBlobs includes all the inputs except first.txt, and
it is only after getting a missing blob error from the remote executor that a second
FindMissingBlobs with all the inputs is sent (followed by a Write of that missing input), and that
the Execute is successful.

This is not something that the client will see externally, but we expect Bazel to include all the
inputs in the initial FindMissingBlobs.

Weirdly, when --disk_cache is not set, Bazel behaves correctly: the first FindMissingBlobs
contains all the inputs, including first.txt.